refactor(aggregate,check-target): split no-baseline audit logic into sibling modules to satisfy architecture/debt.yaml
CI's ai-readiness job failed on commit d6b38238f: the "Architecture
contract (ADR-061)" debt-no-growth gate rejected six files this PR's
cumulative no-baseline-audit fixes had pushed over their recorded
line-count ceiling (abicheck/buildsource/check_report.py,
abicheck/workflows/aggregate/fold.py, abicheck/workflows/aggregate/load.py,
tests/test_action_validate_inputs.py, tests/test_aggregate.py,
tests/test_check_report.py).

Per AGENTS.md's "the way to shrink an entry is to move responsibility out
to a properly-owned module, never to trim the file to fit" -- pure code
moves, no behavior change, verified by running every affected test suite
unchanged:

- abicheck/buildsource/check_report_no_baseline.py (new): the four
  no-baseline-audit-specific pieces of check_report.py's derive_effective_
  depth/_neutralize_gate/_stamp_schema_version/_classify_verdict, mirroring
  the existing check_report_run_outcome.py sibling-split pattern.
- abicheck/workflows/aggregate/no_baseline_load.py (new): the no-baseline
  branch of load.py's _load_report_file, as load_no_baseline_report().
  _LoadedReport itself moved to contracts.py (load.py's own dataclass,
  needed by the new sibling without an import cycle) and
  _analysis_assurance_exit moved to gate.py (next to its sibling axis
  reader _contract_coverage_exit) for the same reason.
- abicheck/workflows/aggregate/profile_matrix_render.py (new): fold.py's
  _render_profile_entry_line, a pure function of one ProfileMatrixEntry,
  moved as a plain module-level function with fold.py's method reduced to
  a one-line delegator.
- tests/test_check_report_no_baseline.py,
  tests/test_action_validate_inputs_audit_only.py (new): the matching test
  classes/methods split out of their parent files, following the
  established tests/test_aggregate_no_baseline_audit.py pattern.
- tests/test_aggregate.py: a dict-comprehension reformatted back to the
  file's own single-line convention (7 lines).

architecture/debt.yaml itself is untouched -- every file is back at or
under its existing adoption baseline, so no baseline needed raising.

mypy/ruff clean; every touched test suite passes unchanged.
